requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

python-dotenv==1.0.1

openai==1.12.0

faiss-cpu==1.7.4

youtube-transcript-api==0.6.2
youtube-search-python==1.6.6

numpy==1.26.4
requests==2.31.0
reportlab
python-pptx